Moved cookies consent view
